Start working on stitch testing
